requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.25
requests==2.25.1
faker==9.13.1
python-telegram-bot==13.6
pandas==1.3.2
PyMySQL==1.0.2
